hw/riscv: Fix integer overflow in cm_base calculation
Cast CM_SIZE to uint64_t before multiplying by the loop counter
to avoid potential integer overflow.

Resolves: Coverity CID 1644076

Message-Id: <20260112090458.2345280-2-djordje.todorovic@htecgroup.com>

+++ hw/riscv/cps.c
@@ -133,7 +133,7 @@ static void riscv_cps_realize(DeviceState *dev, Error **errp)
                             sysbus_mmio_get_region(SYS_BUS_DEVICE(&s->gcr), 0));
 
     for (i = 0; i < num_of_clusters; i++) {
-        uint64_t cm_base = GLOBAL_CM_BASE + (CM_SIZE * i);
+        uint64_t cm_base = GLOBAL_CM_BASE + ((uint64_t)CM_SIZE * i);
         uint32_t hartid_base = i << MHARTID_CLUSTER_SHIFT;
         s->aplic = riscv_aplic_create(cm_base + AIA_PLIC_M_OFFSET,
                                       AIA_PLIC_M_SIZE,
